Add safe pay contract
As we can't verify the addresses that are set in register values for
implementer or donators, they might be incorrect (by mistake or
maliciously). If an incorrect address is written down in contract registers
the raffle may stuck in reward, gift unwrap/redeem and ticket redeem
transactions.

The safe pay contract is added to prevent such scenarios. If an address is
incorrect, the payment to that address will be stuck, and they will
lose their funds. However, raffle procedures will continue properly
without any problem.