Add project ergo tree to context data
SafePay contract only carries the address hash. To facilitate the safe
payment transaction we are going to add the regarding address to the
context data.

In success raffle license redeem transaction, the safe pay box is created
to pay the project fund. Project ergo tree is added to the transaction to
use in safe payment transaction.